💚Adding fixes based on our meeting with TA
# Description

Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.

Author: # Daniel Dickson, Leo Song
Fixes # 1

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
If there were not any tests completed for this, please make sure the comments on your code have insightful information for future reference.
- [X] Ran code and it worked, previous errors were fixed

# Checklist:

- [X] My code follows the style guidelines of this project (PEP 8)
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] My changes generate no new warnings
- [X] Any dependent changes have been merged and published in downstream modules