💚 attempting to fix some errors
# Description

Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.

Author: # Daniel Dickson
Fixes # several

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] This change requires a documentation update

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
If there were not any tests completed for this, please make sure the comments on your code have insightful information for future reference.
- [x] testing via flake8 on github since my computer still won't let me host

# Checklist:

- [x] My code follows the style guidelines of this project (PEP 8)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] Any dependent changes have been merged and published in downstream modules