Create DockerFile
professor Ding: define what is installed and included in the container for our web application. Basically the FROM statement defines the base image, then we just install dependencies and copy the application over the image. We also copy the wait-for-it.sh file there for later use. The EXPOSE 8081 statement means that the container will expose its port 8081 to the outside of the container. So we can visit the app with our host.